// apps/web/src/shared/api/generated/model/seasonHistoryEntry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Gravit API Docs
 * 앱센터 16.5기 동계 프로젝트 Gravit API Docs
 * OpenAPI spec version: 1.0.0
 */

export interface SeasonHistoryEntry {
  seasonKey: string;
  displayKey: string;
  leagueName: string;
  sortOrder: number;
  isCurrent: boolean;
}